Plant Leaf Disease Detection/requirements.txt
setuptools>=68.0.0
wheel>=0.42.0
numpy>=1.26.0
tensorflow>=2.15.0
streamlit==1.31.0
Pillow==10.2.0